+++ gen/template/datastore/bigquery/ddl/schema.sql
@@ -1,0 +1,15 @@
+CREATE OR REPLACE TABLE dummy_type (
+  id       INT64 NOT NULL,
+  name     STRING,
+  modified TIMESTAMP
+);
+
+
+CREATE OR REPLACE TABLE dummy (
+  id         INT64 NOT NULL,
+  type_id    INT64 NOT NULL,
+  name       STRING,
+  modified   TIMESTAMP
+);
+
+